Move svd_approx from decomp.py to lanczos.py and delete decomp.py
